feat: makes CSV times more deterministic
