+++ src/buckets/components/csvUploader/CsvUploaderBody.tsx
@@ -1,18 +1,17 @@
-import React, {FC} from 'react'
+import React, {FC, useContext} from 'react'
 import DragAndDrop from 'src/buckets/components/csvUploader/DragAndDrop'
 import {ProgressBar} from '@influxdata/clockface'
+import {CsvUploaderContext} from 'src/buckets/components/context/csvUploaderProvider'
 
-type Props = {
-  handleDrop: (csv: string) => void
-  hasFile: boolean
-  value: number
+const CsvUploaderBody: FC = () => {
+  const {handleDrop, hasFile, value} = useContext(CsvUploaderContext)
+
+  return (
+    <>
+      <div>{hasFile && <ProgressBar value={value} max={100} />}</div>
+      <DragAndDrop className="line-protocol--content" onSetBody={handleDrop} />
+    </>
+  )
 }
-
-const CsvUploaderBody: FC<Props> = ({handleDrop, hasFile, value}) => (
-  <>
-    <div>{hasFile && <ProgressBar value={value} max={100} />}</div>
-    <DragAndDrop className="line-protocol--content" onSetBody={handleDrop} />
-  </>
-)
 
 export default CsvUploaderBody
