+++ src/buckets/containers/BucketsIndex.tsx
@@ -12,6 +12,7 @@ import GetAssetLimits from 'src/cloud/components/GetAssetLimits'
 import LimitChecker from 'src/cloud/components/LimitChecker'
 import LineProtocolWizard from 'src/buckets/components/lineProtocol/LineProtocolWizard'
 import CsvUploaderWizard from 'src/buckets/components/csvUploader/CsvUploaderWizard'
+import CsvUploaderProvider from 'src/buckets/components/context/csvUploaderProvider'
 import CollectorsWizard from 'src/dataLoaders/components/collectorsWizard/CollectorsWizard'
 import UpdateBucketOverlay from 'src/buckets/components/UpdateBucketOverlay'
 import RenameBucketOverlay from 'src/buckets/components/RenameBucketOverlay'
@@ -60,10 +61,12 @@ class BucketsIndex extends Component {
             component={LineProtocolWizard}
           />
           {isFlagEnabled('csvUploader') && (
-            <Route
-              path={`${bucketsPath}/csv-uploader/new`}
-              component={CsvUploaderWizard}
-            />
+            <CsvUploaderProvider>
+              <Route
+                path={`${bucketsPath}/csv-uploader/new`}
+                component={CsvUploaderWizard}
+              />
+            </CsvUploaderProvider>
           )}
           <Route
             path={`${bucketsPath}/telegrafs/new`}
